requirements.txt
flask==3.0.3
flask-sqlalchemy==2.0.30
pymysql==1.1.1
Flask-WTF==1.2.1